# meta: make NameEquals alloc-free