Benchstat results:
```
name                 old time/op    new time/op    delta
NameEquals/1part-8     76.2ns ± 2%    15.3ns ± 2%   -79.94%  (p=0.008 n=5+5)
NameEquals/3parts-8     174ns ± 1%      30ns ± 1%   -82.74%  (p=0.008 n=5+5)

name                 old alloc/op   new alloc/op   delta
NameEquals/1part-8      16.0B ± 0%      0.0B       -100.00%  (p=0.008 n=5+5)
NameEquals/3parts-8     48.0B ± 0%      0.0B       -100.00%  (p=0.008 n=5+5)

name                 old allocs/op  new allocs/op  delta
NameEquals/1part-8       1.00 ± 0%      0.00       -100.00%  (p=0.008 n=5+5)
NameEquals/3parts-8      1.00 ± 0%      0.00       -100.00%  (p=0.008 n=5+5)
```

Refs #364

Signed-off-by: Iskander Sharipov <quasilyte@gmail.com>